Register Neutron cache options
We currently seem to rely on someone else registering the cache options.
In the rare cases that nobody else registers these options we run into a
NoSuchOptError exception. So, to make sure we have the opts and can use
the cache we now register this ourselves. As I am not sure if we only
need this in the l3 part vs if we need it in the ml2 part either now or
later I decided to just register the ops for everyone in
register_common_opts(). This problem only appears in the agent code so
far.