Delete and readd NatPool for reconfiguration
When the NatPool of a router needs updating (e.g. when the addresses in
the pool are changed), we need to delete and then readd the pool to the
device. If we try to update the pool and there are existing connections
on the pool YANG will error out with a CLI error like this:

%Pool POOL-6435a9a639fe464eb3342511e4e6efff in use, cannot redefine

Deleting the pool will cause all current connections to be dropped, but
this is something that is expected anyway when we update the external
gateway of a router.

It's also worth noting that whenever we see the "cannot redefine" error,
confd will do a full device sync for the /native config tree, which
might again lock the device for quite some time.