Remove ordering code for external interface IPs
`external_fixed_ips` was ordered to become a stable primary IP address
for the interface. This code is not necessary anymore, as if there are
more than 1 fixed IPs on the external interface,
`ASR1KPluginBase._update_router_gw_info()` will record these IPs in the
NAT pool RouterAtts. `GatewayInterface._ip_address()` will then again
use this information and determine the stable primary IP of the
GatwayInterface. Consequently an ordering is not needed at this place
anymore.

This code was also making sure that each supplied fixed IP got a
prefixlen. I am unsure how it would even get at this place, without a
prefixlen, as everything needs to have a subnet which determines the
prefixlen.