[WIP] Prevent uncleaned external interface IP to cause firmware race condition
The current firmware mal-allocates an external interfaces NAT pool (the
pointer to the hash-table, I presume) if there is an external interface
with the same IP in another VRF. We believe this is erroneous behavior
as VRFs are supposed to carry overlapping IP spaces, yet _the vendor_
argues otherwise.

Generally we would like to maintain the support of overlapping NAT
overload IPs at least from a driver perspective as theoretically this
would be valid in OpenStack with multiple adress scopes, never the less
it is unlikely to be seen in our infra.

This commit implements a preflight check, checking if any other outside
interface is present that carries the same IP as our to be programmed
interface. If that is, we must check if that is an uncleaned residue
from a VRF that is still to be cleaned or if it is a valid interface. If
it is not, we delete that interface and program the new interface with
the matching IP, thereby preventing the mal-allocation of the internal
NAT pool.