Resolve diamond shaped inheritance
In the previous implementation `firewall.AccessList` would inherit from
`l3.AccessList` which inherited from `base.Base`.
`firewall.AccessList` also inherited from
`firewall.FirewallPolicyObject`. That lead to a Diamond Shaped
inheritance. The fact that `FirewallPolicyObject` and `l3.AccessList`
had constructors with different arguments made it difficult to let
`super()` call the constructors with different arguments.

To resolve that, the diamond Shaped inheritance was resolved making it
clearer which `__init__`'s `super()` calls which superclasses
constructor.

The different signatures of the constructor still pose a problem, hence
we send all supplemental arguments of a constructor to `**kwargs`.