Migrate to new prefix list yang model
Cisco migrated their previously hierarchical way of managing prefix
lists (one entry per prefix list, then a list of entries per list) to a
model, where every entry is equal to each other, but they're now keyed
by name and no. As this new option is available on 17.6 and the old
model was deprecated and removed with 17.15, we now need to migrate to
the new model.

The new style brings with it some challenges, as we normally handle
prefix lists on a name-basis, but now we need to take care of them by
no/name. To keep the old behavior inside the driver, we consolidate the
seqs into a single prefix list when the config is pulled down from the
device.

We also have a problem with "getting a single item" vs "getting multiple
items", as the methods we're using have different needs here. As all
NyBase.get() methods filter for a single prefix list name, we return
only a single item in all cases where we only find one element.
Otherwise we return multiple items, which for example get_all_stubs()
can handle without problems. It is a bit confusing and if we ever rework
how the driver handles single/multiple instances of a class this would
be one of the first points to start refactoring.