Better prefix-list diff
Previously, we only diffed a prefix list if it had internal interfaces
set, but we might also want to diff a list if it has not. Therefore we
now rely on the `has_prefixes` attribute of a class, set by all
subclasses. If the list has prefixes, diff it. If not, it should be None
on the device.

We were also only ever returning one prefix-list with diffs, as the old
code was overwriting older diffs. The prefix-list entry in the diff dict
is now a list to which we append, so we should now see all the diffs.