Migrate to new dynamic nat model
Cisco changed their config for nat and split the <list> tag into
<list-interface> and <list-pool>. Under these two tags, we now have a
slightly different config format, where a lot of the keys now have a
"-new" in their name. As 17.6 already supports this format, we can
switch the whole driver to the new format.

Details can be found online here:
https://www.yangcatalog.org/yang-search/yang_tree/Cisco-IOS-XE-nat@2022-07-01