Remove support for BDI/hyperloop cable
Before we had BD-VIF/bridge support on our routes we needed to have a
hyperloop cable (traversing two port channels, using up like two thirds
of our router ports). Now that all of this is long behind us we can also
remove the compatibility code for switching back to BDI. This means we
remove the PortChannels LoopbackExternalInterface and
LoopbackInternalInterface and also switch-over support by removing the
bd_iftype (tied to an asr1k context), the use_bdvif flag (both on a
context and in the oslo config part of the driver) and the force_bdi
flag (used to fetch BDIs on a router already on BD-VIFs so we can still
remove them). VBInterface was renamed to BDInterface, same goes for the
related classes VBPrimaryIpAddress, VBSecondaryIpAddress and
VBInterfaceStats.

For migrating the code I updated all conditionals in a way as if
bd_iftype is always BD-VIF and use_bdvif is always True.