IPv6 support for Route Maps
Route Maps can now reference IPv6 prefix lists. We also did some
maintenance on the RouteMap class, like removing the unused enable_bgp
flag and making sure that specifying both prefix_list and access_list
don't override each other's match conditions (though this hasn't been
used so far).

As we don't know yet how we'll transport IPv6 prefixes from ACI/EVPN to
our routers we'll for now leave the RoutePrefix to be v4 only and don't
reference it for v6 addresses. Once we know how that part of the infra
is going to work we'll fill it with life!