Add address scope to subnets in get_sync_data()
With IPv6 we want to make sure that we know the address scope for each
subnet, so we can easily decide in the agent if a network is routable or
not. With IPv4 we normally only have one subnet per port (and I'm not
even sure we can have more than one), but with IPv6 subnets of the same
network are grouped on the same router port.