Reintroduce BDInterface.get_for_vrf()
In the heat of cleaning up the old BDI code and for that matter not
needing to parameterize all the get_item/get_filter calls we also
accidentally removed the get_for_vrf() call. This might result in that
in some cases we don't clean up interfaces properly. This happens on
router delete and the exception is caught, so the DeviceCleaner should
remove any left over interfaces, but reintroducing the original method
is of course the way to go.

By readding get_for_vrf() we no longer see an exception on router delete
and everything should get cleaned up properly.