Use context provided by request in rpc server
In the ml2 mechanism driver we used a pre-created RPC context object
instead of the context object provided by the call itself. That means
that all calls use the same request with the same request id, which
makes tracing calls harder. In one instance this also lead to the DB
connection of this object being broken, which rendered the context
unusable for all subsequent calls. To prevent all of this, we're now
using the context provided by the request.