FWaas: Handle rules with protocol any
If the user specifies protocol any via the FWaaS OpenStack API we end up
with a rule having protocol None. With the old behavior this results in
an empty protocol field, which is not accepted by our hardware router.
Hence, we have to provide a protocol for the ACL. If we don't want to
specify an protocol (tcp, udp, icmp), we need to specify the address
family ("ip") instead, so we now default to that value.

Leaving out the protocol tag from the netconf yang, does not work. It is
accepted by the device (i.e. the YANG stack), but with the empty
protocol field, the YANG stack replaces this with the value "any" (which
I would have done as well if I hadn't read the Cisco device help) and
the IOS-XE cli does not accept any as a valid value here, resulting in a
<bad-cli> error.