Properly remove prefix seqs in preflight()
With the introduction of the new prefix list yang model in c626065 we
also needed to introduce a cleanup method in preflight() that cleans up
stale entries of a prefix list. This removal did not work properly, as
we provided a seq as a string (number), while the Prefix class expects a
PrefixSeq yang model. Therefore the cleanup in preflight failed with:

AttributeError: asr1k_neutron_l3.models.netconf_yang.prefix in
to_delete_dict 'str' object has no attribute 'no'

When we convert the prefix list to a proper model, cleanup works.