IPv6 support for ext iface loop prevention
We want to replicate the loop prevention implemented in
a63bf9432a680a1bebd4a8cd888dee77b02f1c73 for IPv4. This code should
prevent traffic coming from the external interface from leaving the
external interface again, if the vrf decides it needs to leave via the
default route. This can happen when the device the router is connected
to (e.g. ACI) thinks a destination is in a vrf, but the vrf doesn't have
the destination anymore (due to removed fips, removed subnets or
whatever). With IPv4 we had an explicit "set precedence routine" for
internal packets, but when trying to configure this on the device cli it
looks like this is omitted as being the default config. As we don't need
to replicate default config we'll leave that part as-is.

This change results in the following cli change for routers with an ipv6
address on the external interface:

interface BD-VIF 9999
+ ipv6 traffic-filter common ACL-EXT-TOS-V6 out
+ ipv6 policy route-map RM-EXT-TOS-V6

The access-list ACL-EXT-TOS-V6 and route-map RM-EXT-TOS-V6 need to be
pre-provisioned.