Add first sync data tests for db to yang model
We now have the first few tests that can create a router in the
OpenStack database and generate sync data from this, which can then be
used to instantiate the models used in the agent that then generate the
netconf yang config for our devices. This means we can more or less now
do end-to-end tests for DB-to-YANG tests, including coverage for if the
Router() class correctly handles the data gotten from the DB and makes
the right decisions.

The current mock-implementation is a best effort approach to get us
started with having tests. As the sync data is normally transferred via
RPC / rabbitmq and then augmented with the address scope info we need to
run some extra steps in the tests in get_sync_data() to have everything
similar to what we have in production.