Improve VRF AddressFamily handling
We want to be able to add and remove the address-family of a VRF as
needed, so now we only enable it, if we also have an ip address for it -
at least in the case of IPv6. For IPv4 we - for now - always keep the AF
enabled to make sure we don't have any problems with the config
transition. Once we are at a point were we make all device config
dependant on the enable_ipv4 flag we can also think about removing the
AF from teh VRF.

We have a special case in the netconf-yang pulled from the device: an
empty address-family config results in an empty tag, e.g. <ipv6/>, which
then results in this config not appearing in the diff. A diff is needed
for the object to actually be updated, so we add an extra flag to the
Vrf netconf class, that creates an empty address family object if we
find an empty tag on the device.