Clean up PoolDynamicNat on orphan VRF delete
When a router is deleted, we clean up all objects relating to a VRF
before deleting the VRF itself on the device. If this fails, we have
the device cleaner cleaning up the orphaned objects, but for this to
work we have to also make sure depending objects are deleted. This is
already done for InterfaceDynamicNat, but not for PoolDynamicNat - so we
now do it for PoolDynamicNat as well.