Make IPv4 fips syncable on IPv6-only routers
Routers with an IPv6-only gateway can have IPv4 fips. Before this
commit, we were passing the unused global ip mask to the StaticNat
class, which required us to have an IPv4 gateway address. This is not
needed and should make this kind of router syncable again.

I am not yet fully sure if IPv4 fips will work on this kind of router,
but a sync error is definitely not the way to go there. Either we get
this working or we forbid it in a precommit hook.