Delete empty prefix lists on extisting routers
Since the prefix list data model has been changed, we can't just delete
"a prefix list" anymore, we always have to delete all the entries. This
is easy when we know which entries should be in the prefix list. If we
want to do a replace, we need to first delete everything that should not
be in the list by fetching these entries from the device itself and then
add what should be on the device. This is done by our preflight hook.

With the code before this commit we only call the preflight hook if
there are still entries left in the list, but now, for existing routers,
we want to always call this preflight hook. Therefore we're now removing
the explicit call of delete() on empty prefix list.