Better dynamic NAT pool gateway allocation logic
For dynamic NAT pooling we always have one block of consecutive IPv4
addresses and one gateway address, which does not need to be part of the
block. If the gateway IP is not specified by the user we left it to
Neutron to allocate a fitting IP, which randomly allocates one. This
caused problems for users as this random IP might fragment up the IP
space of their network, resulting in less routers with NAT pools fitting
in the network.

To fix this we now use a more deterministic approach: We allocate the
highest available IP in the network. As we always try to use the
smallest block with the lowest starting IP for the NAT pool we should
end up having less fragmentations in the IP space. Other methods we
thought of were using first available after pool or first available
around pool but this might interfere with other pool allocations in case
we have many routers entering and leaving the network (it also might
not, I don't have any reliable data on this - the decision is based on
educated guesses).

We also now retry the allocation in the rare case that in between
selecting and allocating the IP someone else allocated it, with up to 5
tries and a random wait between 0 and 0.05s. The retry is only done when
an IpAddressAlreadyAllocated exception occurs. This might also lead to a
retry if the user specifies already allocated options, leading to a few
more DB queries and a slower query of 0.2s.

For retrying we use tenacity, which is also used by neutron. The asr1k
driver already uses retrying, but this library is no longer maintained
and we should switch to tenacity for everything else in the foreseeable
future.

The allocation logic for the NAT pool itself now also changes slightly,
in that we make sure that we sort all available IP address space chunks
not only by size, but also by their first IP. So in cases where we have
multiple chunks of the same size we should end up with the chunk with
the smallest starting IP.