Use router flavors for scheduling
Agents now have traits and routers can be created with a router flavor,
which can have traits as well. We support required and optional traits,
which control which router can be scheduled on which agent. This
information is only used for scheduling.

For the router traits are part of the flavor. A flavor can have one or
multiple service profiles. Each service profile has a metainfo field,
which is a text field, where we try to parse json out of. A valid
config would be a dictionary with one or both of the keys req_traits and
opt_traits, containing a list of strings, which represent the traits a
router could or should have.  If the trait information of a flavor
service profile is invalid or has the wrong type, the information will
be discarded and a warning will be logged.

On the agent side, this is done via agent config and then exposed to
neutron-server via the agent configuration submitted with the agent's
state.

Scheduling is then done by checking which router and agent combination
has all the required traits on each side.

We don't support antiaffinity yet ("this router should NOT be scheduled
on a router that has the 100g tag"), but this could easily be
implemented if we ever need it. The "unschedulable" feature that we
currently handle via an extra parameter is now also part of the traits,
but we currently submit it as trait and config for migration purposes
(can be removed once this has been deployed once). We could also handle
the AZ parts via the traits feature inside the scheduler by adding it to
the routers/agents required traits when scheduling to make the
scheduling code less complex, but we'll lave that for a future commit.

With the flavor framework we need to register our own ASR1KRouterDriver,
so we can have flavor service profiles with a valid driver. The
configuration should look something like this:

[service_providers]
service_provider = L3_ROUTER_NAT:asr1k:asr1k_neutron_l3.neutron.services.service_providers.asr1k_router.ASR1KRouterDriver:default