Dynamic quota registration for flavors
We now have the ability to register and enforce quota for certain
flavors.

All network flavors that have a service profile attached which specifies
'"req_quota": true' as part of their json dict metadata will now get
registered as quota. The quota will be prefixed with "routers_flavor_",
meaning that a flavor with the name "high-throughput" would be exposed
as a quota named "routers_flavor_high-throughput". Enforcement of the
quota is done on router create. Flavors are refreshed periodically (by
default every 15 minutes), on start of neutron-server or when a flavor
is used that requires quota. Quotas are implemented as a
CountableResource, which doesn't track anything itself, but looks up how
many routers of a project have the relevant flavor attached and returns
that number.