Add garp feature flag, disable it by default
The garp feature is not yet compatible with the no-alias feature. To
proceed with the first rollout of 17.15 we're now introducing a feature
switch for garp and leave it off by default. Once we're ready for
deploying garp everywhere we'll either change the default, add a version
detection flag or configure it on a per-agent basis (but it's most
likely going to be autodetection).

We are also removing the garp flag else-case which originally should
have removed the garp flag from config.  As StaticNatList is already
called with operation="replace" we don't have to do an explicit removal
(and if we would use operation="remove" for the garp flag we would in
fact get an error saying "The operation remove is not allowed under
replace").