Implement VPNaaS API for IOS XE
Implement the Neutron VPNaaS API backend to support IPSec on Cisco XE
devices. This is a large commit, as most of the moving parts needed to
implement this only make sense together.

Architecture: A VPNaaS router in our infrastructure is not allowed to
have any subnets. Subnets need to be connected via BGPVPN. As we don't
know the local subnets, the user needs to specify them all via an
endpoint group. To make sure routers are scheduled on our VPNaaS
hardware, the vpnaas trait is required for them (and also required for
attaching a vpn service object to a router).

VPNaaS driver: We have a VPNaas service driver, much like our FWaaS and
BGPVPN service driver. It takes care of most of the validation via
registry hooks. In some places this replaces the direct calls used
normally by VPNaaS, as validation needs to happen inside the DB
transaction or needs to fetch some related data before some object is
deleted. In most cases we then notify the agent to do a router sync,
except for ipsec site connection deletion (see section on deletion
below). VPNaaS information is fetched as part of the get_sync_data() and
is then incorporated into the models.neutron.l3.Router class.

API extensions: For our VPNs we need configurable internal tunnel ips
(IPv4 and IPv6) and the ability to configure a NAT IP. Both these things
need to be configured by the user, so we provide API extensions for
these two cases. Each has their own api_def, which extends the
ipsec-site-connection resource.

Internal tunnel IPs: These IPs, provided by an extra extension, are used
for each ipsec site connection that we have. If they are not provided by
the user they are auto allocated from the link local pool of the address
family. We do some validation to make sure the IPs of site connections
on the same router do not overlap. The user can specify a local subnet
and a peer address. If the peer address is not part of the local subnet,
then a static route is added for this.

Tunnel ids: A Tunnel interface on Cisco hardware needs a number. As we
need to remember this number, we have a table in our db for it. The
number gets allocated from a configurable range on a per-agent basis and
is deleted once the ipsec site connection is deleted.

Neutron VPNaaS: We depend on our own brand of VPNaaS[0], as we need some
extra notifications (DB precommit/postcommit hooks) and resource
extender for extra API fields. Additionally, the "disconnected subnet
mode" config is needed for router-subnet validation.

Endpoint groups: We only allow cidr endpoint groups, as our VPNaaS
routers don't have directly connected subnets. We impose a maximum of
100 endpoints (by default, value is configurable) per endpoint group, as
in transport mode we need to create an ACL with source x destination
entries and decided 10000 is a good upper limit for ACL rule count.

Device/YANG: We require Cisco IOS XE >= 17.15, else some YANG objects
are not compatible. We configure: IPSecProfile, IPSecTransformSet,
IKEv2Profile, IKEv2Keyring, IKEv2Policy, IKEv2Proposal and
TunnelInterfaces. TunnelInterfaces bring up some YANG problems again, as
deleting a TunnelInterface causes a full sync of the device. This
might cause some problems when deleting / de-scheduling router and also
with the device cleaner.

Device cleaner: Some objects have shared custody (are both referenced
via vrf and a site connection / tunnel id). Objects related to a Tunnel
interface cannot be deleted as long as the Tunnel interface is up, so it
is one of the first objects that we need to delete.

IPSec site connection deletion: We don't clear all related objects when
a site connection is deleted, because a router object only
syncs/overwrites its current state and when a site connection is
deleted there is no longer state for this object. To make sure the
customer tunnel goes down, on site connection deletion we trigger a
Tunnel interface deletion on the device. To delete the Tunnel interface
we also need to delete all objects that depend on it, which currently
are only static routes which have this interface as a dependency. For
all other site connection objects on the device we trust on the device
cleaner to remove them in the next cleaning cycle. They also don't block
us and do not keep interesting state, so we can just leave them there
until the next cleaning.

Router / vpn service / ipsec site connection status: VPN Service and
ipsec site connection have a "status" field, which is set to
PENDING_CREATE by default. Our router only has ACTIVE/ERROR as state,
but this is updated whenever the router is synced by a regular update
(not by an API router sync). For VPNaaS we piggyback onto this mechanism
and set all related VPNaaS objects to ACTIVE when a router has been
synced successfully. For the admin-triggered API sync we also now set
the router status.

IPv6: Support for IPv6 is partly done. It has not been tested and there
are some open questions regarding how we map the OpenStack data model to
Cisco config when we cannot fully offer dual stack. For example, tunnel
mode can only be either ipv4 or ipv6 when the underlay is ipv6, while
with ipv4 underlay a dual stack overlay is allowed (by Cisco config).
This has not been fully mapped out and will be handled in a follow up.

[0] https://github.com/sapcc/neutron-vpnaas/tree/stable/2024.1-m3