Sort routes by prefix + netmask for better diff
Whenever we update something on the device, we check if the config is
already present by doing a diff (an operation also used by admins to
check if the config has been synced properly to the hardware router). In
some cases we now have the same config on the device and in OpenStack,
but the ordering is wrong (as Cisco and Neutron have their separate
ordering). In this case we have this problem for static routes, where
routes with different netmasks are sorted differently.